[IMP] general: merge all docs about unsplash into one doc
This commit is part of a larger work to improve the documentation's
structure. The toctree structure used to add levels to create categories
and make the structure visually more understandable.
With the new version of the documentation, plenty of these categories
are now redundant.

This commit merges all docs about unsplash into a single doc:
- "unsplash/unsplash_access_key" -> unsplash
- "unsplash/unsplash_application_id" -> unsplash

task-2535899