[MOV] general: renames "odoo_basics" into "users" + split content
This commit is part of a larger work to improve the documentation's
structure. The toctree structure used to add levels to create categories
and make the structure visually more understandable.
With the new version of the documentation, plenty of these categories
are now redundant.

Since other docs were moved away from the section "basics" with previous
commits, this category is renamed "users settings"
This commit:
- renames "basics" into "users settings"
- moves the content from "odoo_basics/users" to "users"
  - it moves the content about "access rights" to a child doc

task-2535899